Update README.md
This 15 Puzzle Game in C is a classic sliding puzzle game that uses arrow keys to move numbered tiles on a 4x4 board. The game starts with randomized tile positions, generated by time.h library. Code::Blocks IDE is used to develop and test the game. A rules window is included to guide players before starting the game.